RFM: Let the user handle executables
These automatic copies restrict how the sandbox can be used, for
instance if you happen to play around with several executables that
correspond to the same spec.

Still in case a spack package is used, the corresponding spec must be
provided in order to load the correct environment before submitting
jobs.