REF: modularize user settings further
Implement default namelists in dedicated functions that are git
tracked. If one wants to change a paramter that isn't specified by an
environment variable, it is still possible to redefine and export the
namelist function in a file that has higher priority (user_settings >
config > default). This way the configuration is fully customizable
without the need to modify the `run` scripts which are git tracked.