refactor(chain): move date switching back to job script
This is safer since the number of the part (6_) doesn't matter
anymore as it's replaced by an explicitly declared depdendency. So
that the order of the parts in SB_PARTS doesn't matter anymore. In the
previous implementation, if 6_chain appeared before the end, it would
have swithed the dates for parts of the current chunk.